+++ src/components.ts
@@ -14,4 +14,5 @@ export * from './file-card';
 export * from './folder';
 export * from './code-highlighter';
 export * from './mermaid';
+export * from './x-markdown';
 export * from './notification';